# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nabd HMS API
  version: "1.0"
  description: >
    Foundation-layer contract (Auth, RBAC/Staff, Tenant/Branch, Patients, Audit).
    See README.md for conventions. Deny-by-default authorization: every
    operation below declares x-required-permission; the gateway rejects
    anything not explicitly granted.
servers:
  - url: https://api.nabd.health/v1
    description: production
  - url: https://api.staging.nabd.health/v1
    description: staging
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: "PIN login (step 1 of 2 when MFA is enforced). NFR-06: no SSO, no hardware keys, no password — PIN/WhatsApp-OTP/authenticator app are the only three methods."
      security: []
      x-required-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, email, pin]
              properties:
                tenantSlug: { type: string, description: "Clinic subdomain/code; disambiguates the same email across two tenants" }
                email: { type: string, format: email }
                pin: { type: string, pattern: "^[0-9]{4,6}$" }
      responses:
        "200":
          description: PIN accepted; MFA challenge issued if enforced for this role
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TokenPair"
                  - $ref: "#/components/schemas/MfaChallenge"
        "401":
          description: Invalid credentials (identical timing/body for unknown vs. wrong-PIN — no user enumeration)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
        "423":
          description: Account locked (exponential backoff in effect)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  /auth/otp/request:
    post:
      tags: [Auth]
      summary: "WhatsApp OTP login, step 1 — sends a 6-digit code. ponytail: delivery is mocked/logged until the Messaging Platform epic (E17) integrates the real WhatsApp Business API."
      security: []
      x-required-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, mobilePhone]
              properties:
                tenantSlug: { type: string }
                mobilePhone: { type: string }
      responses:
        "202": { description: "Always 202 regardless of whether the tenant/mobile is known — no enumeration" }
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: WhatsApp OTP login, step 2 — also marks the staff member's mobile as verified (NB-041)
      security: []
      x-required-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, mobilePhone, code]
              properties:
                tenantSlug: { type: string }
                mobilePhone: { type: string }
                code: { type: string, pattern: "^[0-9]{6}$" }
      responses:
        "200":
          description: TokenPair, or an MfaChallenge if MFA is enforced for this role
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TokenPair"
                  - $ref: "#/components/schemas/MfaChallenge"
        "401":
          description: Wrong/expired code or unknown tenant+mobile (identical shape — no enumeration)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
        "423":
          description: Account locked (exponential backoff in effect)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Complete login with TOTP/SMS code, or mint a short-lived step-up token on an existing session
      security:
        - bearerAuth: []
        - {}
      x-required-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code]
              properties:
                challengeId: { type: string, description: "Opaque signed token, not a UUID" }
                code: { type: string, pattern: "^[0-9]{6}$" }
      responses:
        "200":
          description: TokenPair (login flow) or a step-up token (already-authenticated flow)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TokenPair"
                  - $ref: "#/components/schemas/StepUpToken"
        "401":
          description: Wrong code, expired challenge, or 5th failed attempt (challenge burned)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access/refresh pair
      security: []
      x-required-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New pair issued; old refresh token invalidated
          content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } }
        "401":
          description: Reused/revoked refresh token — entire session family revoked as a compromise signal
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  /auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Remotely revoke a specific session (e.g. lost device)
      x-required-permission: authenticated
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204": { description: "Revoked; denied within 60s per NB-032" }
        "404": { $ref: "#/components/responses/NotFound" }

  # NB-046 (PIN reset & account recovery): specced, not yet implemented in the backend.
...
  /patients:
    get:
      tags: [Patients]
      summary: List/search patients (row-scoped — a doctor with "own patients only" sees only their own)
      x-required-permission: patients:view
      parameters:
        - name: q
          in: query
          description: Free-text match on name (any script), phone, MRN or token
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/PatientPage" }
    post:
      tags: [Patients]
      summary: Register a patient
      x-required-permission: patients:create
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatientWrite" } } }
      responses:
        "201": { description: "Patient", content: { application/json: { schema: { $ref: "#/components/schemas/Patient" } } } }
        "409":
          description: Likely-duplicate match found (phone/name/DOB fuzzy match) — client should offer merge instead of blind create
          content: { application/json: { schema: { type: object, properties: { candidates: { type: array, items: { $ref: "#/components/schemas/PatientMatchCandidate" } } } } } }
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Patient 360 (identity, allergies, active packages, balance, last visit)
      x-required-permission: patients:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "PatientDetail", content: { application/json: { schema: { $ref: "#/components/schemas/PatientDetail" } } } }
        "404":
          description: >
            Not found OR exists but out of the caller's row-level scope
            (e.g. another doctor's patient) — always 404, never 403, so an
            ID probe can't be used to confirm a patient exists (NB-041).
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    patch:
      tags: [Patients]
      x-required-permission: patients:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatientWrite" } } }
      responses:
        "200": { description: "Patient", content: { application/json: { schema: { $ref: "#/components/schemas/Patient" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /patients/{id}/merge:
    post:
      tags: [Patients]
      summary: Merge a duplicate into this patient (supervisor action, reversible for 30 days)
      x-required-permission: patients:delete
      x-step-up: true
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [duplicatePatientId]
              properties:
                duplicatePatientId: { type: string, format: uuid }
      responses:
        "200": { description: "PatientDetail", content: { application/json: { schema: { $ref: "#/components/schemas/PatientDetail" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /patients/merges/{mergeId}/undo:
    post:
      tags: [Patients]
      summary: Reverse a merge — only within its 30-day window (NB-072)
      x-required-permission: patients:delete
      x-step-up: true
      parameters:
        - name: mergeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Reversed }
        "404": { description: "Merge doesn't exist, already reversed, or past the 30-day window", content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /patients/{id}/consent/withdraw:
    post:
      tags: [Patients]
      summary: Withdraw a consent — removes the patient from every consent-gated read path immediately (NB-053)
      x-required-permission: patients:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consentType: { type: string, enum: [treatment, data_processing, messaging], default: data_processing }
      responses:
        "204": { description: Withdrawn }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------- Queue
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token. Carries tenantId, staffId, roleId and permission set as
        claims — the server trusts only these, never client-supplied
        tenant/role headers.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    NotFound:
      description: Not found (or out of scope — see per-endpoint note)
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    PatientPage:
      description: Page of patients
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: { $ref: "#/components/schemas/Patient" } }
              page: { $ref: "#/components/schemas/PageMeta" }
  schemas:
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri, example: "https://api.nabd.health/errors/validation-failed" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties:
        nextCursor: { type: [string, "null"] }
        limit: { type: integer }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 900 }
    MfaChallenge:
      type: object
      properties:
        challengeId: { type: string, description: "Opaque signed token, not a UUID" }
        method: { type: string, enum: [totp, sms] }
        expiresIn: { type: integer, example: 300 }
    StepUpToken:
      type: object
      properties:
        stepUpToken: { type: string }
        expiresIn: { type: integer, example: 300 }
...
    PatientWrite:
      type: object
      required: [name, phone, dob, gender]
      properties:
        name: { type: string, description: "Stored/searched as-entered; any script" }
        phone: { type: string }
        dob: { type: string, format: date }
        gender: { type: string, enum: [male, female, other] }
        guardianId: { type: [string, "null"], format: uuid, description: "Required if patient is a minor" }
        address: { type: string }
        nationalId:
          type: [string, "null"]
          description: "Aadhaar/Iqama — column-level encrypted at rest (NB-140), never returned in list responses"
    Patient:
      type: object
      properties:
        id: { type: string, format: uuid }
        mrn: { type: string }
        name: { type: string }
        phone: { type: string }
        dob: { type: string, format: date }
        gender: { type: string, enum: [male, female, other] }
        status: { type: string, enum: [active, merged] }
    PatientDetail:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            allergies: { type: array, items: { type: string } }
            chronicConditions: { type: array, items: { type: string } }
            activePackages: { type: integer }
            outstandingBalance: { type: number }
            lastVisitAt: { type: [string, "null"], format: date-time }
    PatientMatchCandidate:
      type: object
      properties:
        patientId: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        matchScore: { type: number }
...